refactor(core): share shadow shade-time (ShadowMap / PCF / PCSS) via maquette-core
The shadow shade path — the ShadowMap depth map with project / lit / lit_pcss,
the LightShadow dispatch, and BiasParams — was byte-identical between the mesh
and core; only BiasParams precision differed cosmetically (both already compute
the depth test in f64). Unified BiasParams to f64, added a ShadowMap::new
constructor + a public splat_tri, and had the mesh re-export the shared shade
half from core.

The frustum *builder* stays per-consumer on purpose: the mesh frames the light
view at the scene centre (a robust approximation), while gltf aims along the
light's own direction with the real spot cone — a legitimate requirement
difference, not accidental duplication.

Renders are byte-identical for BOTH consumers (verified on a shadowed teapot and
tokyo). mesh shadow.rs: 405 -> 110 lines.

+++ crates/maquette-gltf/src/render.rs
@@ -160,7 +160,7 @@ fn rasterize_scene(buffer: &mut PixelBuffer, scene: &Scene, config: &RenderConfi
             &caster_tris, &raw_lights, bc, effective_br, up, sh_cfg.resolution,
         );
         let bias = maquette_core::shadow::BiasParams {
-            bias: sh_cfg.bias, normal_bias: sh_cfg.normal_bias, slope_bias: sh_cfg.slope_bias,
+            bias: sh_cfg.bias as f64, normal_bias: sh_cfg.normal_bias as f64, slope_bias: sh_cfg.slope_bias as f64,
         };
         (maps, bias, sh_cfg.softness, sh_cfg.pcss_light_size)
     } else {
